requirements.txt
google-generativeai
youtube-transcript-api
google-api-python-client
pandas
python-dotenv